Moving kin.ts into extensions/kinemage
